+++ MC-CBS/MC-CBS/general-graph/LLNode.cpp
@@ -7,7 +7,7 @@
 
 LLNode::LLNode() : g_val(0), h_val(0), parent(NULL), timestep(0), num_internal_conf(0), in_openlist(false) {}
 
-LLNode::LLNode(const vertex_t& vertex, int g_val, int h_val, LLNode* parent, int timestep, int num_internal_conf, bool in_openlist) :
+LLNode::LLNode(const vertex_t& vertex, double g_val, double h_val, LLNode* parent, int timestep, int num_internal_conf, bool in_openlist) :
 	vertex(vertex), g_val(g_val), h_val(h_val), parent(parent), timestep(timestep),
 	num_internal_conf(num_internal_conf), in_openlist(in_openlist) {}
